refactor: move transform to separate module
